Move RbsrStore to wgps crate
