+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 
-members = ["data-model", "earthstar", "encoding", "fuzz", "meadowcap"]
+members = ["data-model", "earthstar", "encoding", "fuzz", "meadowcap", "wgps"]
 resolver = "2"
 
 [workspace.lints.clippy]
