+++ wgps/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "wgps"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+willow-data-model = { path = "../data-model", version = "0.1.0" }
+
+[lints]
+workspace = true
